+++ db/schema.rb
@@ -63,7 +63,7 @@ ActiveRecord::Schema.define(version: 2021_01_03_145140) do
     t.string "municipalities", null: false
     t.string "address", null: false
     t.string "building_name"
-    t.integer "phone_number", null: false
+    t.bigint "phone_number", null: false
     t.bigint "order_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
